form changes
I made changes in my js, SCSS and html for my style.
In my stylesheet, I placed all contact form style in one place because it was in different places. I added a comment to it. In html, I got rid of all things that prevented my JavaScript from running. In my js I commented out some code to see if the rest of it works. I still got some errors with it that need to be resolved.