changes in form
I made a bug changes in my form again. in my html, I added  "onkeyup" to my inputs, I replaced divs error messages with spans error messages. I also added success icons to my js function that I wrote again and now everything work as expected. The only problem is that the button automatically refreshes the page but it's not the issue if JavaScript, the issue is somewhere else. Need resolving. Also styling for error messages and success icons need to be added.