form - some more changes
I removed one of form's field and replaced the other one with "full name". I did the same in my JavaScript. I removed one of variables and one function. I changed the other function, added a regex for full mane, modified some manes and texts inside that function and in variable's name. I resolved the issue with the submit button that was refreshing the page and prevented the last function from running.