Regex for name validation, asterisks added to the form
I replaced the name regex with a different one and it eliminated the white spaces issues. I modified it a little bit to allow users use a name that is shorter than three characters.
I also added asterisks to the form texts that shows what fields are required. I added them in my html to reach label and styled them with my CSS