minor changes
I modified my JavaScript function for the form validation and made the form to be cleared after submitting. I added the addEventListener() after reset that displays a success message which is displayed for 3 seconds. I modified my html and placed the success message in a div outside the form. I also removed a typo that I found in one of my texts.