HTML and CSS "Coding Examples" page update
I updated the "Coding Example" page. I replaced the js code picture with the updated HTML <pre> and  the <code> with its class of the code language name and placed my js code inside <code>. I used the Prism.js library to highlight the comment, string, number and operator and then I used each of them as CSS class and added a colour to each of them.
I also updated the code description.